+++ CHANGES.md
@@ -1,6 +1,6 @@
 2.1.1 (unreleased)
 ------------------
-- 
+- Remove trailing garbage from output lines
 
 2.1.0 (2018-08-25)
 ------------------